==> data_collection/01_03-data_collection-twitch_streamer_stats.qmd <==
---
title: "Mining: Twitch streamer statistics"
execute: 
  eval: false
---


```{r}
# Create a tibble based on the provided data
library(tidyverse)

data_hasanabi <- tribble(
  ~Month, ~Avg_Viewers, ~Avg_Viewers_Gain, ~Avg_Viewers_Percent_Gain, ~Peak_Viewers, ~Hours_Streamed, ~Hours_Streamed_Gain, ~Hours_Streamed_Percent_Gain, ~Followers, ~Followers_Gain, ~Followers_Percent_Gain, ~Followers_Per_Hour,
  "Dec 2024", 26265, -17031, -39.3, 37094, 18.4, -221, -92.3, 2793538, 604, NA, 32.8,
  "Nov 2024", 43296, 12510, 40.6, 312431, 239, 16.8, 7.6, 2792934, 59296, 2.2, 248,
  "Oct 2024", 30786, 2921, 10.5, 110889, 223, -12.9, -5.5, 2733638, 17998, 0.7, 80.9,
  "Sep 2024", 27865, -8174, -22.7, 192633, 236, 3.9, 1.7, 2715640, 26341, 1, 112,
  "Aug 2024", 36039, -57, -0.2, 86598, 232, 12.1, 5.5, 2689299, 23519, 0.9, 102,
  "Jul 2024", 36096, 14830, 69.7, 135265, 220, 3.7, 1.7, 2665780, 37035, 1.4, 169,
  "Jun 2024", 21266, -1291, -5.7, 126069, 216, -9.5, -4.2, 2628745, 4805, 0.2, 22.3,
  "May 2024", 22557, 1939, 9.4, 37656, 225, -24.4, -9.8, 2623940, 61255, 2.4, 272,
  "Apr 2024", 20618, 1980, 10.6, 64992, 250, 31, 14.2, 2562685, 4545, 0.2, 18.2,
  "Mar 2024", 18638, -374, -2, 46536, 219, -8.5, -3.7, 2558140, -4783, -0.2, -21.9,
  "Feb 2024", 19012, -714, -3.6, 34541, 227, 8.3, 3.8, 2562923, 690, NA, 3,
  "Jan 2024", 19726, 1504, 8.3, 37875, 219, 31.5, 16.8, 2562233, 1035, NA, 4.7
  ) %>% 
  janitor::clean_names() %>% 
  mutate(streamer = "hasanabi")
```

```{r}
# Create a tibble based on the provided data
data_zackrawrr <- tribble(
  ~Month, ~Avg_Viewers, ~Avg_Viewers_Gain, ~Avg_Viewers_Percent_Gain, ~Peak_Viewers, ~Hours_Streamed, ~Hours_Streamed_Gain, ~Hours_Streamed_Percent_Gain, ~Followers, ~Followers_Gain, ~Followers_Percent_Gain, ~Followers_Per_Hour,
  "Dec 2024", 31240, -17585, -36, 41061, 13.5, -155, -92, 1914946, 280, NA, 20.8,
  "Nov 2024", 48825, 11527, 30.9, 170847, 168, 57.9, 52.5, 1914666, 21419, 1.1, 127,
  "Oct 2024", 37298, 5369, 16.8, 80561, 110, -93.5, -45.9, 1893247, 9610, 0.5, 87.2,
  "Sep 2024", 31929, -1415, -4.2, 71902, 204, 8.8, 4.5, 1883637, 21098, 1.1, 104,
  "Aug 2024", 33344, 1054, 3.3, 74397, 195, -5.1, -2.5, 1862539, 22585, 1.2, 116,
  "Jul 2024", 32290, 4919, 18, 93783, 200, 14.4, 7.8, 1839954, 25887, 1.4, 130,
  "Jun 2024", 27371, 10911, 66.3, 62836, 186, -55.8, -23.1, 1814067, 25388, 1.4, 137,
  "May 2024", 16460, -1888, -10.3, 48105, 241, 39.2, 19.4, 1788679, 15092, 0.9, 62.5,
  "Apr 2024", 18348, -2500, -12, 35976, 202, 12.7, 6.7, 1773587, 18650, 1.1, 92.3,
  "Mar 2024", 20848, 1142, 5.8, 40612, 190, 24.4, 14.7, 1754937, 32400, 1.9, 171,
  "Feb 2024", 19706, -5189, -20.8, 41302, 165, 0.8, 0.5, 1722537, 36598, 2.2, 222,
  "Jan 2024", 24895, -6579, -20.9, 45829, 164, -29.2, -15.1, 1685939, 42503, 2.6, 259
  ) %>% 
  janitor::clean_names() %>% 
  mutate(streamer = "zackrawrr")

```


```{r}
# Create a tibble based on the provided data
data_tmr <- tribble(
  ~Month, ~Avg_Viewers, ~Avg_Viewers_Gain, ~Avg_Viewers_Percent_Gain, ~Peak_Viewers, ~Hours_Streamed, ~Hours_Streamed_Gain, ~Hours_Streamed_Percent_Gain, ~Followers, ~Followers_Gain, ~Followers_Percent_Gain, ~Followers_Per_Hour,
  "Dec 2024", 1005, -632, -38.6, 1971, 7.1, -55.2, -88.6, 83325, 200, 0.2, 28.1,
  "Nov 2024", 1637, 588, 56.1, 4995, 62.3, -8.2, -11.6, 83125, 5639, 7.3, 90.5,
  "Oct 2024", 1049, 102, 10.8, 2383, 70.5, -0.2, -0.3, 77486, 2098, 2.8, 29.8,
  "Sep 2024", 947, -131, -12.2, 2144, 70.7, -1.6, -2.2, 75388, 1075, 1.4, 15.2,
  "Aug 2024", 1078, 197, 22.4, 2189, 72.3, -0.7, -0.9, 74313, 2215, 3.1, 30.6,
  "Jul 2024", 881, 242, 37.9, 1971, 73, 10, 15.9, 72098, 2058, 2.9, 28.2,
  "Jun 2024", 639, 7, 1.1, 1230, 62.9, -3.7, -5.5, 70040, 691, 1, 11,
  "May 2024", 632, 83, 15.1, 1100, 66.6, -1.5, -2.2, 69349, 715, 1, 10.7,
  "Apr 2024", 549, -52, -8.7, 905, 68.1, 4.2, 6.6, 68634, 501, 0.7, 7.4,
  "Mar 2024", 601, 71, 13.4, 1206, 63.9, -1.3, -2, 68133, 427, 0.6, 6.7,
  "Feb 2024", 530, -22, -4, 1907, 65.2, 0.2, 0.3, 67706, 156, 0.2, 2.4,
  "Jan 2024", 552, 33, 6.4, 971, 65, 9.6, 17.3, 67550, 232, 0.3, 3.6
  ) %>% 
  janitor::clean_names() %>% 
  mutate(streamer = "the_majority_report")
```



```{r}
data <- bind_rows(data_hasanabi, data_zackrawrr, data_tmr) %>%
  mutate(month = lubridate::my(month))
```

```{r}
qs::qsave(data, file = "local_data/twitch_streamer_stats.qs")
```
